added status panel that can use item in overworld + dialogue

+++ Assets/Scripts/Attacks/BaseAttack.cs
@@ -7,6 +7,6 @@ public class BaseAttack : MonoBehaviour
 {
     public string attackName;
     public string attackDescription;
-    public float attackDamage;
-    public float attackCost;
+    public int attackDamage;
+    public int attackCost;
 }
